requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
docopt==0.6.2
Flask==2.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pipreqs==0.4.11
protobuf==3.19.4
pycparser==2.21
PyJWT==2.4.0
requests==2.28.1
six==1.16.0
urllib3==1.26.10
waitress==2.1.2
Werkzeug==2.1.2
yarg==0.1.9
zipp==3.8.1
